refactor SchemaRegistryConfig construction
